Move functions to the adequate files
