+++ frontend/src/sagas/groupSaga.js
@@ -1,5 +1,10 @@
 import { call, put, takeEvery, all, fork } from "redux-saga/effects";
-import { showGroupReq, editGroupReq } from "../api/groupRequests";
+import {
+  showGroupReq,
+  editGroupReq,
+  deleteGroupReq,
+  createGroupReq,
+} from "../api/groupRequests";
 import {
   showGroupRequest,
   showGroupSuccess,
@@ -7,6 +12,12 @@ import {
   editGroupRequest,
   editGroupSuccess,
   editGroupError,
+  deleteGroupRequest,
+  deleteGroupSuccess,
+  deleteGroupError,
+  createGroupRequest,
+  createGroupSuccess,
+  createGroupError,
 } from "../slices/groupSlice";
 import { snackbarNotification } from "../helpers/notifierHelper";
 import { showGroupsRequest } from "../slices/groupsSlice";
@@ -56,6 +67,58 @@ export function* watcherEditGroup() {
   yield takeEvery(editGroupRequest, workerEditGroup);
 }
 
+export function* workerDeleteGroup({ payload }) {
+  try {
+    console.log(payload);
+    const { group_id, target_group } = payload;
+    let res = yield call(deleteGroupReq, group_id, target_group);
+    yield put(deleteGroupSuccess(res.data));
+    yield put(showGroupsRequest());
+  } catch (error) {
+    yield put(deleteGroupError());
+    yield put(snackbarNotification("Failed to delete group.", "error"));
+  }
+}
+
+export function* watcherDeleteGroup() {
+  yield takeEvery(deleteGroupRequest, workerDeleteGroup);
+}
+
+export function* workerCreateGroup({ payload }) {
+  try {
+    console.log(payload);
+    const {
+      groupId,
+      groupName,
+      password,
+      hasAdminPrivileges,
+      accessibleDemos,
+    } = payload;
+    let res = yield call(
+      createGroupReq,
+      groupId,
+      groupName,
+      password,
+      hasAdminPrivileges,
+      accessibleDemos
+    );
+    yield put(createGroupSuccess(res.data));
+    yield put(showGroupsRequest());
+  } catch (error) {
+    yield put(createGroupError());
+    yield put(snackbarNotification("Failed to create group.", "error"));
+  }
+}
+
+export function* watcherCreateGroup() {
+  yield takeEvery(createGroupRequest, workerCreateGroup);
+}
+
 export default function* rootSaga() {
-  yield all([fork(watcherShowGroups), fork(watcherEditGroup)]);
+  yield all([
+    fork(watcherShowGroups),
+    fork(watcherEditGroup),
+    fork(watcherDeleteGroup),
+    fork(watcherCreateGroup),
+  ]);
 }
