+++ frontend/src/sagas/groupsSaga.js
@@ -1,19 +1,9 @@
 import { call, put, takeEvery, all, fork } from "redux-saga/effects";
-import {
-  showGroupsReq,
-  createGroupReq,
-  deleteGroupReq,
-} from "../api/groupsRequests";
+import { showGroupsReq } from "../api/groupsRequests";
 import {
   showGroupsRequest,
   showGroupsSuccess,
   showGroupsError,
-  createGroupRequest,
-  createGroupSuccess,
-  createGroupError,
-  deleteGroupRequest,
-  deleteGroupSuccess,
-  deleteGroupError,
 } from "../slices/groupsSlice";
 import { snackbarNotification } from "../helpers/notifierHelper";
 
@@ -31,57 +21,6 @@ export function* watcherShowGroups() {
   yield takeEvery(showGroupsRequest, workerShowGroups);
 }
 
-export function* workerDeleteGroup({ payload }) {
-  try {
-    console.log(payload);
-    const { group_id, target_group } = payload;
-    let res = yield call(deleteGroupReq, group_id, target_group);
-    yield put(deleteGroupSuccess(res.data));
-    yield put(showGroupsRequest());
-  } catch (error) {
-    yield put(deleteGroupError());
-    yield put(snackbarNotification("Failed to delete group.", "error"));
-  }
-}
-
-export function* watcherDeleteGroup() {
-  yield takeEvery(deleteGroupRequest, workerDeleteGroup);
-}
-
-export function* workerCreateGroup({ payload }) {
-  try {
-    console.log(payload);
-    const {
-      groupId,
-      groupName,
-      password,
-      hasAdminPrivileges,
-      accessibleDemos,
-    } = payload;
-    let res = yield call(
-      createGroupReq,
-      groupId,
-      groupName,
-      password,
-      hasAdminPrivileges,
-      accessibleDemos
-    );
-    yield put(createGroupSuccess(res.data));
-    yield put(showGroupsRequest());
-  } catch (error) {
-    yield put(createGroupError());
-    yield put(snackbarNotification("Failed to create group.", "error"));
-  }
-}
-
-export function* watcherCreateGroup() {
-  yield takeEvery(createGroupRequest, workerCreateGroup);
-}
-
 export default function* rootSaga() {
-  yield all([
-    fork(watcherShowGroups),
-    fork(watcherDeleteGroup),
-    fork(watcherCreateGroup),
-  ]);
+  yield all([fork(watcherShowGroups)]);
 }
