+++ frontend/src/slices/groupSlice.js
@@ -30,6 +30,24 @@ export const groupSlice = createSlice({
     editGroupError: (state, action) => {
       state.isEditingGroup = false;
     },
+    createGroupRequest: (state, action) => {
+      state.isCreatingGroup = true;
+    },
+    createGroupSuccess: (state, action) => {
+      state.isCreatingGroup = false;
+    },
+    createGroupError: (state, action) => {
+      state.isCreatingGroup = false;
+    },
+    deleteGroupRequest: (state, action) => {
+      state.isDeletingGroup = true;
+    },
+    deleteGroupSuccess: (state, action) => {
+      state.isDeletingGroup = false;
+    },
+    deleteGroupError: (state, action) => {
+      state.isDeletingGroup = false;
+    },
     resetGroupState: (state, action) => initialState,
   },
 });
@@ -40,6 +58,12 @@ export const {
   editGroupRequest,
   editGroupSuccess,
   editGroupError,
+  createGroupRequest,
+  createGroupSuccess,
+  createGroupError,
+  deleteGroupRequest,
+  deleteGroupSuccess,
+  deleteGroupError,
   resetGroupState,
 } = groupSlice.actions;
